+++ __tests__/setupTests.ts
@@ -1,0 +1,5 @@
+//mock custom-card-helpers
+jest.mock('custom-card-helpers', () => ({
+  ...jest.requireActual('custom-card-helpers'),
+  stateIcon: jest.fn().mockReturnValue('state-icon'),
+}));