+++ rollup.config.dev.js
@@ -22,7 +22,7 @@ export default {
     serve({
       contentBase: "./dist",
       host: "0.0.0.0",
-      port: 5000,
+      port: 3000,
       allowCrossOrigin: true,
       headers: {
         "Access-Control-Allow-Origin": "*",