+++ src/editor/index.ts
@@ -183,9 +183,9 @@ export class SankeyChartEditor extends LitElement implements LovelaceCardEditor 
         type: 'grid',
         name: '',
         schema: [
-          { name: 'min_state', selector: { number: { mode: 'box' } } },
+          { name: 'min_state', selector: { number: { mode: 'box', min: 0., step: 'any' } } },
+          { name: 'static_scale', selector: { number: { mode: 'box' } } },
           { name: 'round', selector: { number: { mode: 'box', unit_of_measurement: localize('editor.decimals') } } },
-          { name: 'throttle', selector: { number: { mode: 'box', unit_of_measurement: 'ms' } } },
           {
             name: 'unit_prefix',
             selector: {
@@ -195,6 +195,29 @@ export class SankeyChartEditor extends LitElement implements LovelaceCardEditor 
               },
             },
           },
+          {
+            name: 'sort_by',
+            selector: {
+              select: {
+                mode: 'dropdown',
+                options: [{ value: 'none', label: localize('editor.sort_by.none') }, { value: 'state', label: localize('editor.sort_by.state') }],
+              },
+            },
+          },
+          {
+            name: 'sort_dir',
+            selector: {
+              select: {
+                mode: 'dropdown',
+                options: [
+                  { value: '' },
+                  { value: 'desc', label: localize('editor.sort_dir.desc') },
+                  { value: 'asc', label: localize('editor.sort_dir.asc') },
+                ],
+              },
+            },
+          },
+          { name: 'throttle', selector: { number: { mode: 'box', unit_of_measurement: 'ms' } } },
         ],
       },
     ];
@@ -209,7 +232,8 @@ export class SankeyChartEditor extends LitElement implements LovelaceCardEditor 
       return html``;
     }
 
-    const config = normalizeConfig(this._config || ({} as SankeyChartConfig));
+    const isMetric = this.hass.config.unit_system.length == "km";
+    const config = normalizeConfig(this._config || ({} as SankeyChartConfig), isMetric);
     const { autoconfig } = config;
     const sections: SectionConfig[] = config.sections || [];