+++ src/utils.ts
@@ -1,12 +1,13 @@
 import { createThing, formatNumber, FrontendLocaleData, HomeAssistant, LovelaceCard, LovelaceCardConfig } from 'custom-card-helpers';
 import { html, TemplateResult } from 'lit';
-import { DEFAULT_ENTITY_CONF, UNIT_PREFIXES } from './const';
+import { DEFAULT_ENTITY_CONF, UNIT_PREFIXES, FT3_PER_M3 } from './const';
 import {
   Box,
   ChildConfigOrStr,
   Config,
   Connection,
   ConnectionState,
+  EntityConfigInternal,
   EntityConfigOrStr,
   SankeyChartConfig,
   Section,
@@ -17,7 +18,7 @@ export function cloneObj<T extends Record<string, unknown>>(obj: T): T {
   return JSON.parse(JSON.stringify(obj));
 }
 
-export function formatState(state: number, round: number, locale: FrontendLocaleData): string {
+export function formatState(state: number, round: number, locale: FrontendLocaleData, monetary_unit?: string): string {
   let rounded: string;
   let decimals = round;
   do {
@@ -25,7 +26,15 @@ export function formatState(state: number, round: number, locale: FrontendLocale
     rounded = state.toFixed(decimals++);
   } while (/^[0\.]*$/.test(rounded) && decimals < 100);
 
-  return formatNumber(parseFloat(rounded), locale);
+  if (!monetary_unit) {
+    return formatNumber(parseFloat(rounded), locale);
+  } else {
+    return formatNumber(parseFloat(rounded), locale, {
+      style: "currency",
+      currency: monetary_unit,
+      minimumFractionDigits: round,
+    });
+  }
 }
 
 export function normalizeStateValue(
@@ -34,7 +43,7 @@ export function normalizeStateValue(
   unit_of_measurement?: string,
 ): { state: number; unit_of_measurement?: string } {
   const validState = Math.max(0, state);
-  if (!unit_of_measurement) {
+  if (!unit_of_measurement || unit_of_measurement == 'monetary') {
     return { state: validState, unit_of_measurement };
   }
   const prefix = Object.keys(UNIT_PREFIXES).find(p => unit_of_measurement!.indexOf(p) === 0) || '';
@@ -53,15 +62,21 @@ export function getEntityId(entity: EntityConfigOrStr | ChildConfigOrStr): strin
   return typeof entity === 'string' ? entity : entity.entity_id;
 }
 
-export function getChildConnections(parent: Box, children: Box[], connections?: ConnectionState[]): Connection[] {
+export function getChildConnections(parent: Box, children: Box[], allConnections: ConnectionState[], connectionsByParent: Map<EntityConfigInternal, ConnectionState[]>): Connection[] {
   // @NOTE don't take prevParentState from connection because it is different
   let prevParentState = 0;
+  let state = 0;
+  const childConnections = connectionsByParent.get(parent.config);
   return children.map(child => {
-    const connection = connections?.find(c => c.child.entity_id === child.entity_id);
-    if (!connection) {
-      throw new Error(`Missing connection: ${parent.entity_id} - ${child.entity_id}`);
+    let connections = childConnections?.filter(c => c.child.entity_id === child.entity_id);
+    if (!connections?.length) {
+      connections = allConnections
+        .filter(c => c.passthroughs.includes(child) || c.passthroughs.includes(parent.config));
+      if (!connections.length) {
+        throw new Error(`Missing connection: ${parent.entity_id} - ${child.entity_id}`);
+      }
     }
-    const { state, prevChildState } = connection;
+    state = connections.reduce((sum, c) => sum + c.state, 0);
     if (state <= 0) {
       // only continue if this connection will be rendered
       return { state } as Connection;
@@ -69,8 +84,10 @@ export function getChildConnections(parent: Box, children: Box[], connections?: 
     const startY = (prevParentState / parent.state) * parent.size + parent.top;
     prevParentState += state;
     const startSize = Math.max((state / parent.state) * parent.size, 0);
-    const endY = (prevChildState / child.state) * child.size + child.top;
+    const endY = (child.connectedParentState / child.state) * child.size + child.top;
     const endSize = Math.max((state / child.state) * child.size, 0);
+
+    child.connectedParentState += state;
 
     return {
       startY,
@@ -80,12 +97,12 @@ export function getChildConnections(parent: Box, children: Box[], connections?: 
       endSize,
       endColor: child.color,
       state,
-      highlighted: connection.highlighted,
+      highlighted: connections.some(c => c.highlighted),
     };
   });
 }
 
-export function normalizeConfig(conf: SankeyChartConfig): Config {
+export function normalizeConfig(conf: SankeyChartConfig, isMetric: boolean): Config {
   let config = { sections: [], ...cloneObj(conf) };
 
   const { autoconfig } = conf;
@@ -169,10 +186,15 @@ export function normalizeConfig(conf: SankeyChartConfig): Config {
     });
   });
 
+  const default_co2_per_ft3 = 55.0  // gCO2e/ft3 tailpipe
+                            + 11.6; // gCO2e/ft3 supply chain, US average
   return {
     height: 200,
     unit_prefix: '',
     round: 0,
+    convert_units_to: '',
+    co2_intensity_entity: 'sensor.co2_signal_co2_intensity',
+    gas_co2_intensity: isMetric ? default_co2_per_ft3 * FT3_PER_M3 : default_co2_per_ft3,
     // handle legacy min_box_height
     min_box_size: config.min_box_height ?? 3,
     min_box_distance: 5,